feat: remove alunos and pessoais features, add production environment configuration
