+++ back/src/main/resources/application.properties
@@ -4,7 +4,10 @@ spring.datasource.url=${DB_URL}
 spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASSWORD}
 spring.jpa.hibernate.ddl-auto=none
-spring.jpa.show-sql=true
+spring.jpa.show-sql=${SHOW_SQL:false}
+
+# CORS
+cors.allowed-origins=${CORS_ALLOWED_ORIGINS:http://localhost:4200,http://127.0.0.1:4200}
 spring.messages.basename=messages
 spring.messages.encoding=UTF-8
 logging.level.org.flywaydb=DEBUG
