+++ back/src/main/resources/db/migration/V21__create_revoked_token.sql
@@ -1,3 +1,5 @@
+-- V20 foi intencionalmente pulada. A próxima migration deve ser V24__.
+
 CREATE TABLE IF NOT EXISTS revoked_token (
     jti        VARCHAR(255) PRIMARY KEY,
     expires_at TIMESTAMP    NOT NULL
