+++ frontend/repz/angular.json
@@ -25,7 +25,9 @@
             "tsConfig": "tsconfig.app.json",
             "inlineStyleLanguage": "scss",
             "stylePreprocessorOptions": {
-              "includePaths": ["src/styles"]
+              "includePaths": [
+                "src/styles"
+              ]
             },
             "assets": [
               {
@@ -39,13 +41,18 @@
               }
             ],
             "styles": [
-                "node_modules/primeicons/primeicons.css",
-                "src/styles.scss"
+              "node_modules/primeicons/primeicons.css",
+              "src/styles.scss"
             ],
             "server": "src/main.server.ts",
             "outputMode": "server",
             "security": {
-              "allowedHosts": ["localhost", "127.0.0.1", "0.0.0.0", "repzfront"]
+              "allowedHosts": [
+                "localhost",
+                "127.0.0.1",
+                "0.0.0.0",
+                "repzfront"
+              ]
             },
             "ssr": {
               "entry": "src/server.ts"
@@ -59,13 +66,19 @@
                   "maximumWarning": "500kB",
                   "maximumError": "1MB"
                 },
-               {
-                "type": "anyComponentStyle",
-                "maximumWarning": "10kb",
-                "maximumError": "12kb"
-              }
+                {
+                  "type": "anyComponentStyle",
+                  "maximumWarning": "10kb",
+                  "maximumError": "12kb"
+                }
               ],
-              "outputHashing": "all"
+              "outputHashing": "all",
+              "fileReplacements": [
+                {
+                  "replace": "src/environments/environment.ts",
+                  "with": "src/environments/environment.prod.ts"
+                }
+              ]
             },
             "development": {
               "optimization": false,
@@ -78,7 +91,12 @@
         "serve": {
           "builder": "@angular/build:dev-server",
           "options": {
-            "allowedHosts": ["localhost", "127.0.0.1", "0.0.0.0", "repzfront"]
+            "allowedHosts": [
+              "localhost",
+              "127.0.0.1",
+              "0.0.0.0",
+              "repzfront"
+            ]
           },
           "configurations": {
             "production": {
