--- frontend/repz/src/app/features/alunos/alunos.html
@@ -1,230 +1,0 @@
-<app-shell
-  ativo="alunos"
-  [crumbs]="[((isAdmin() ? 'ROLES.ADMIN' : 'ROLES.MANAGER') | translate), ('NAV.STUDENTS' | translate)]"
-  [nome]="isAdmin() ? 'Administrador' : 'Academia'"
-  [subtitulo]="(isAdmin() ? 'ADMIN.AREA' : 'ACADEMIA.AREA') | translate"
-  [mostrarCta]="false"
->
-  <p-toast />
-  <p-confirmDialog />
-
-  <div class="page-header">
-    <div>
-      <div class="page-title">{{ 'NAV.STUDENTS' | translate }}</div>
-      <div class="page-subtitle">
-        {{ (isAdmin() ? 'MGMT.STUDENTS.SUBTITLE_ADMIN' : 'MGMT.STUDENTS.SUBTITLE_GYM') | translate }}
-      </div>
-    </div>
-    <div class="header-actions">
-      <p-select
-        *ngIf="isAdmin()"
-        [options]="academias()"
-        optionLabel="name"
-        optionValue="id"
-        [(ngModel)]="academiaSelecionada"
-        (onChange)="trocarAcademia()"
-        [placeholder]="'MGMT.STUDENTS.SELECT_GYM' | translate"
-        [style]="{ 'min-width': '240px' }"
-      />
-      <p-button
-        [label]="'MGMT.STUDENTS.NEW_STUDENT' | translate"
-        icon="pi pi-plus"
-        styleClass="repz-primary"
-        (onClick)="novo()"
-        [disabled]="isAdmin() && !academiaSelecionada()"
-      />
-    </div>
-  </div>
-
-  <p-card styleClass="repz-card state-card" *ngIf="carregando()">
-    <p-progressSpinner strokeWidth="3" animationDuration=".8s" styleClass="state-spinner" />
-    <span>{{ 'MGMT.STUDENTS.LOADING' | translate }}</span>
-  </p-card>
-
-  <p-card styleClass="repz-card state-card state-error" *ngIf="!carregando() && erro()">
-    <i class="pi pi-exclamation-triangle"></i>
-    <span>{{ erro() }}</span>
-    <p-button [label]="'ADMIN.GYMS.RETRY' | translate" icon="pi pi-refresh" (onClick)="carregar()" />
-  </p-card>
-
-  <p-card styleClass="repz-card" *ngIf="!carregando() && !erro()">
-    <div class="hist-head">
-      <div class="font-semibold">{{ 'MGMT.STUDENTS.REGISTERED' | translate }}</div>
-      <div class="text-3 text-sm">{{ 'ALUNO.AVAL.N_RECORDS' | translate: { n: lista().length } }}</div>
-    </div>
-
-    <div class="empty-message text-3 text-sm" *ngIf="lista().length === 0">
-      {{ 'MGMT.STUDENTS.NONE_FOUND' | translate }}
-    </div>
-
-    <p-table
-      *ngIf="lista().length > 0"
-      [value]="lista()"
-      styleClass="repz-table"
-      [tableStyle]="{ 'min-width': '900px' }"
-      [paginator]="lista().length > 10"
-      [rows]="10"
-    >
-      <ng-template pTemplate="header">
-        <tr>
-          <th>{{ 'PROFILE.NAME' | translate }}</th>
-          <th>{{ 'PROFILE.EMAIL' | translate }}</th>
-          <th>{{ 'PROFILE.PLAN' | translate }}</th>
-          <th>{{ 'ROLES.TRAINER' | translate }}</th>
-          <th *ngIf="isAdmin()">{{ 'PROFILE.GYM' | translate }}</th>
-          <th>{{ 'PERSONAL.STUDENTS.COL_STATUS' | translate }}</th>
-          <th style="width: 160px"></th>
-        </tr>
-      </ng-template>
-      <ng-template pTemplate="body" let-a>
-        <tr>
-          <td>
-            <strong>{{ a.nome }}</strong>
-            <div class="text-3 text-sm" *ngIf="a.objetivo">{{ a.objetivo }}</div>
-          </td>
-          <td>{{ a.email }}</td>
-          <td>{{ a.planoNome || '—' }}</td>
-          <td>{{ a.personalNome || '—' }}</td>
-          <td *ngIf="isAdmin()">{{ a.academiaNome || '—' }}</td>
-          <td>
-            <p-tag
-              [value]="(a.ativo ? 'COMMON.ACTIVE' : 'COMMON.INACTIVE') | translate"
-              [severity]="a.ativo ? 'success' : 'danger'"
-            />
-          </td>
-          <td>
-            <div class="actions-cell">
-              <p-button icon="pi pi-pencil" text rounded severity="secondary" (onClick)="editar(a)" />
-              <p-button
-                *ngIf="a.ativo"
-                icon="pi pi-ban"
-                text
-                rounded
-                severity="danger"
-                (onClick)="confirmarInativar(a)"
-              />
-              <p-button
-                *ngIf="!a.ativo"
-                icon="pi pi-check"
-                text
-                rounded
-                severity="success"
-                (onClick)="confirmarReativar(a)"
-              />
-            </div>
-          </td>
-        </tr>
-      </ng-template>
-    </p-table>
-  </p-card>
-
-  <p-dialog
-    [(visible)]="dialogAberto"
-    [modal]="true"
-    [closable]="!salvando()"
-    [draggable]="false"
-    [resizable]="false"
-    [style]="{ width: '600px' }"
-    [header]="(modo() === 'editar' ? 'MGMT.STUDENTS.EDIT_STUDENT' : 'MGMT.STUDENTS.NEW_STUDENT') | translate"
-    (onHide)="fecharDialog()"
-  >
-    <form #f="ngForm" (ngSubmit)="salvar(f)" class="form-aluno">
-      <ng-container *ngIf="modo() === 'criar'">
-        <label class="field">
-          <span>{{ 'ADMIN.GYMS.NAME_REQ' | translate }}</span>
-          <input pInputText name="name" [(ngModel)]="form().name" required maxlength="120" />
-        </label>
-        <div class="grid2">
-          <label class="field">
-            <span>{{ 'ACADEMIA.STUDENTS.EMAIL_REQ' | translate }}</span>
-            <input
-              pInputText
-              type="email"
-              name="email"
-              [(ngModel)]="form().email"
-              required
-              email
-            />
-          </label>
-          <label class="field">
-            <span>{{ 'MGMT.STUDENTS.PASSWORD_REQ' | translate }}</span>
-            <input
-              pInputText
-              type="password"
-              name="password"
-              [(ngModel)]="form().password"
-              required
-              minlength="5"
-            />
-          </label>
-        </div>
-      </ng-container>
-
-      <ng-container *ngIf="modo() === 'editar'">
-        <div class="readonly-block">
-          <div><strong>{{ form().name }}</strong></div>
-          <div class="text-3 text-sm">{{ form().email }}</div>
-        </div>
-      </ng-container>
-
-      <div class="grid2">
-        <label class="field">
-          <span>{{ 'PROFILE.PLAN' | translate }} {{ modo() === 'criar' ? '*' : '' }}</span>
-          <p-select
-            name="planoId"
-            [options]="planos()"
-            optionLabel="nome"
-            optionValue="id"
-            [(ngModel)]="form().planoId"
-            [placeholder]="'MGMT.STUDENTS.SELECT' | translate"
-            appendTo="body"
-          />
-        </label>
-        <label class="field">
-          <span>{{ 'ROLES.TRAINER' | translate }}</span>
-          <p-select
-            name="personalId"
-            [options]="personais()"
-            optionLabel="userName"
-            optionValue="id"
-            [(ngModel)]="form().personalId"
-            [placeholder]="'ACADEMIA.STUDENTS.NO_TRAINER' | translate"
-            [showClear]="true"
-            appendTo="body"
-          />
-        </label>
-      </div>
-
-      <label class="field">
-        <span>{{ 'PROFILE.GOAL' | translate }}</span>
-        <input pInputText name="objetivo" [(ngModel)]="form().objetivo" maxlength="160" />
-      </label>
-
-      <p-message
-        *ngIf="erroForm()"
-        severity="error"
-        styleClass="aviso"
-        [text]="erroForm() ?? ''"
-      />
-
-      <div class="form-actions">
-        <p-button
-          type="button"
-          [label]="'COMMON.CANCEL' | translate"
-          severity="secondary"
-          text
-          [disabled]="salvando()"
-          (onClick)="fecharDialog()"
-        />
-        <p-button
-          type="submit"
-          [label]="(modo() === 'editar' ? 'COMMON.SAVE' : 'ADMIN.GYMS.CREATE') | translate"
-          icon="pi pi-save"
-          [loading]="salvando()"
-          [disabled]="salvando()"
-          styleClass="repz-primary"
-        />
-      </div>
-    </form>
-  </p-dialog>
-</app-shell>
