--- frontend/repz/src/app/features/personais/personais.html
@@ -1,194 +1,0 @@
-<app-shell
-  ativo="personais"
-  [crumbs]="[((isAdmin() ? 'ROLES.ADMIN' : 'ROLES.MANAGER') | translate), ('NAV.TRAINERS' | translate)]"
-  [nome]="isAdmin() ? 'Administrador' : 'Academia'"
-  [subtitulo]="(isAdmin() ? 'ADMIN.AREA' : 'ACADEMIA.AREA') | translate"
-  [mostrarCta]="false"
->
-  <p-toast />
-  <p-confirmDialog />
-
-  <div class="page-header">
-    <div>
-      <div class="page-title">{{ 'NAV.TRAINERS' | translate }}</div>
-      <div class="page-subtitle">
-        {{ (isAdmin() ? 'MGMT.TRAINERS.SUBTITLE_ADMIN' : 'MGMT.TRAINERS.SUBTITLE_GYM') | translate }}
-      </div>
-    </div>
-    <div class="header-actions">
-      <p-select
-        *ngIf="isAdmin()"
-        [options]="academias()"
-        optionLabel="name"
-        optionValue="id"
-        [(ngModel)]="academiaSelecionada"
-        (onChange)="trocarAcademia()"
-        [placeholder]="'MGMT.STUDENTS.SELECT_GYM' | translate"
-        [style]="{ 'min-width': '240px' }"
-      />
-      <p-button
-        [label]="'MGMT.TRAINERS.NEW_TRAINER' | translate"
-        icon="pi pi-plus"
-        styleClass="repz-primary"
-        (onClick)="novo()"
-        [disabled]="isAdmin() && !academiaSelecionada()"
-      />
-    </div>
-  </div>
-
-  <p-card styleClass="repz-card state-card" *ngIf="carregando()">
-    <p-progressSpinner strokeWidth="3" animationDuration=".8s" styleClass="state-spinner" />
-    <span>{{ 'MGMT.TRAINERS.LOADING' | translate }}</span>
-  </p-card>
-
-  <p-card styleClass="repz-card state-card state-error" *ngIf="!carregando() && erro()">
-    <i class="pi pi-exclamation-triangle"></i>
-    <span>{{ erro() }}</span>
-    <p-button [label]="'ADMIN.GYMS.RETRY' | translate" icon="pi pi-refresh" (onClick)="carregar()" />
-  </p-card>
-
-  <p-card styleClass="repz-card" *ngIf="!carregando() && !erro()">
-    <div class="hist-head">
-      <div class="font-semibold">{{ 'MGMT.TRAINERS.REGISTERED' | translate }}</div>
-      <div class="text-3 text-sm">{{ 'ALUNO.AVAL.N_RECORDS' | translate: { n: lista().length } }}</div>
-    </div>
-
-    <div class="empty-message text-3 text-sm" *ngIf="lista().length === 0">
-      {{ 'MGMT.TRAINERS.NONE_FOUND' | translate }}
-    </div>
-
-    <p-table
-      *ngIf="lista().length > 0"
-      [value]="lista()"
-      styleClass="repz-table"
-      [tableStyle]="{ 'min-width': '760px' }"
-      [paginator]="lista().length > 10"
-      [rows]="10"
-    >
-      <ng-template pTemplate="header">
-        <tr>
-          <th>{{ 'PROFILE.NAME' | translate }}</th>
-          <th>{{ 'PROFILE.EMAIL' | translate }}</th>
-          <th>{{ 'PROFILE.SPECIALTY' | translate }}</th>
-          <th *ngIf="isAdmin()">{{ 'PROFILE.GYM' | translate }}</th>
-          <th>{{ 'PERSONAL.STUDENTS.COL_STATUS' | translate }}</th>
-          <th style="width: 160px"></th>
-        </tr>
-      </ng-template>
-      <ng-template pTemplate="body" let-p>
-        <tr>
-          <td><strong>{{ p.userName }}</strong></td>
-          <td>{{ p.email }}</td>
-          <td>{{ p.especialidade || '—' }}</td>
-          <td *ngIf="isAdmin()">{{ p.academiaNome || '—' }}</td>
-          <td>
-            <p-tag
-              [value]="(p.ativo ? 'COMMON.ACTIVE' : 'COMMON.INACTIVE') | translate"
-              [severity]="p.ativo ? 'success' : 'danger'"
-            />
-          </td>
-          <td>
-            <div class="actions-cell">
-              <p-button icon="pi pi-pencil" text rounded severity="secondary" (onClick)="editar(p)" />
-              <p-button
-                [icon]="p.ativo ? 'pi pi-ban' : 'pi pi-check'"
-                text
-                rounded
-                [severity]="p.ativo ? 'danger' : 'success'"
-                (onClick)="confirmarToggleAtivo(p)"
-              />
-            </div>
-          </td>
-        </tr>
-      </ng-template>
-    </p-table>
-  </p-card>
-
-  <p-dialog
-    [(visible)]="dialogAberto"
-    [modal]="true"
-    [closable]="!salvando()"
-    [draggable]="false"
-    [resizable]="false"
-    [style]="{ width: '560px' }"
-    [header]="(modo() === 'editar' ? 'MGMT.TRAINERS.EDIT_TRAINER' : 'MGMT.TRAINERS.NEW_TRAINER') | translate"
-    (onHide)="fecharDialog()"
-  >
-    <form #f="ngForm" (ngSubmit)="salvar(f)" class="form-personal">
-      <ng-container *ngIf="modo() === 'criar'">
-        <label class="field">
-          <span>{{ 'ADMIN.GYMS.NAME_REQ' | translate }}</span>
-          <input pInputText name="name" [(ngModel)]="form().name" required maxlength="120" />
-        </label>
-        <div class="grid2">
-          <label class="field">
-            <span>{{ 'ACADEMIA.STUDENTS.EMAIL_REQ' | translate }}</span>
-            <input
-              pInputText
-              type="email"
-              name="email"
-              [(ngModel)]="form().email"
-              required
-              email
-            />
-          </label>
-          <label class="field">
-            <span>{{ 'MGMT.STUDENTS.PASSWORD_REQ' | translate }}</span>
-            <input
-              pInputText
-              type="password"
-              name="password"
-              [(ngModel)]="form().password"
-              required
-              minlength="5"
-            />
-          </label>
-        </div>
-      </ng-container>
-
-      <ng-container *ngIf="modo() === 'editar'">
-        <div class="readonly-block">
-          <div><strong>{{ form().name }}</strong></div>
-          <div class="text-3 text-sm">{{ form().email }}</div>
-        </div>
-      </ng-container>
-
-      <label class="field">
-        <span>{{ 'PROFILE.SPECIALTY' | translate }} {{ modo() === 'criar' ? '' : '*' }}</span>
-        <input
-          pInputText
-          name="especialidade"
-          [(ngModel)]="form().especialidade"
-          [required]="modo() === 'editar'"
-          maxlength="80"
-        />
-      </label>
-
-      <p-message
-        *ngIf="erroForm()"
-        severity="error"
-        styleClass="aviso"
-        [text]="erroForm() ?? ''"
-      />
-
-      <div class="form-actions">
-        <p-button
-          type="button"
-          [label]="'COMMON.CANCEL' | translate"
-          severity="secondary"
-          text
-          [disabled]="salvando()"
-          (onClick)="fecharDialog()"
-        />
-        <p-button
-          type="submit"
-          [label]="(modo() === 'editar' ? 'COMMON.SAVE' : 'ADMIN.GYMS.CREATE') | translate"
-          icon="pi pi-save"
-          [loading]="salvando()"
-          [disabled]="salvando()"
-          styleClass="repz-primary"
-        />
-      </div>
-    </form>
-  </p-dialog>
-</app-shell>
